feat: (admin) add document status
backend: catch parsing error and update document status accordingly
show document status in admin frontend
added redis

+++ admin-backend/src/admin_backend/apis/admin_api.py
@@ -15,6 +15,7 @@ from fastapi import (  # noqa: F401
     Depends,
     Form,
     Header,
+    HTTPException,
     Path,
     Query,
     Request,
@@ -26,6 +27,7 @@ from fastapi import (  # noqa: F401
 )
 
 from admin_backend.models.extra_models import TokenModel  # noqa: F401
+from admin_backend.models.document_status import DocumentStatus
 
 
 router = APIRouter()
@@ -47,7 +49,9 @@ for _, name, _ in pkgutil.iter_modules(ns_pkg.__path__, ns_pkg.__name__ + "."):
 async def delete_document(
     id: str = Path(..., description=""),
 ) -> None:
-    return BaseAdminApi.subclasses[0]().delete_document(id)
+    if not BaseAdminApi.subclasses:
+        raise HTTPException(status_code=500, detail="Not implemented")
+    return await BaseAdminApi.subclasses[0]().delete_document(id)
 
 
 @router.get(
@@ -64,20 +68,24 @@ async def delete_document(
 async def document_reference_id_get(
     id: str = Path(..., description="Identifier of the pdf document."),
 ) -> Response:
-    return BaseAdminApi.subclasses[0]().document_reference_id_get(id)
+    if not BaseAdminApi.subclasses:
+        raise HTTPException(status_code=500, detail="Not implemented")
+    return await BaseAdminApi.subclasses[0]().document_reference_id_get(id)
 
 
 @router.get(
     "/all_documents",
     responses={
-        200: {"model": List[str], "description": "List of document links"},
+        200: {"model": List[DocumentStatus], "description": "List of document links"},
         500: {"description": "Internal server error"},
     },
     tags=["admin"],
     response_model_by_alias=True,
 )
-async def get_all_documents() -> List[str]:
-    return BaseAdminApi.subclasses[0]().get_all_documents()
+async def get_all_documents() -> List[DocumentStatus]:
+    if not BaseAdminApi.subclasses:
+        raise HTTPException(status_code=500, detail="Not implemented")
+    return await BaseAdminApi.subclasses[0]().get_all_documents()
 
 
 @router.post(
@@ -96,4 +104,6 @@ async def upload_documents_post(
     background_tasks: BackgroundTasks,
 ) -> None:
     """Uploads user selected pdf documents."""
+    if not BaseAdminApi.subclasses:
+        raise HTTPException(status_code=500, detail="Not implemented")
     return await BaseAdminApi.subclasses[0]().upload_documents_post(body, request, background_tasks)
